[refactor] rename dao to mapper
