+++ src/main/java/com/wor/dash/config/DBConfig.java
@@ -4,7 +4,7 @@ import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@MapperScan(basePackages={"com.wor.dash.*.model.dao"})
+@MapperScan(basePackages={"com.wor.dash.*.model.mapper"})
 public class DBConfig {
 
 }
