+++ src/main/java/com/wor/dash/jwt/model/service/TokenServiceImp.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.wor.dash.jwt.model.Token;
-import com.wor.dash.jwt.model.dao.TokenMapper;
+import com.wor.dash.jwt.model.mapper.TokenMapper;
 
 import lombok.extern.slf4j.Slf4j;
